+++ CHANGES.rst
@@ -35,6 +35,8 @@ Improvements
 - Show only active registration counts on the registration form management dashboard, and add
   an inactive registration count to the registration list (:pr:`5990`)
 - Store creation date of users and show it to admins (:pr:`5957`, thanks :user:`vasantvohra`)
+- Add option to hide links to Room Booking system for users who lack access (:pr:`5981`,
+  thanks :user:`SegiNyn`)
 
 Bugfixes
 ^^^^^^^^